backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.0
requests==2.31.0
aiofiles==23.2.1
aiohttp==3.8.6

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1

# 异步任务队列
redis==5.0.1
celery==5.3.4

# AI服务
openai==1.2.4
# openai-whisper==20231117
anthropic==0.7.0
Pillow==10.1.0

# 语音识别
ffmpeg-python==0.2.0
torch==2.1.1
torchaudio==2.1.1

# 日志
loguru==0.7.2

# 工具
python-slugify==8.0.1
python-dateutil==2.8.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 开发工具
black==23.10.1
flake8==6.1.0 

# PDF generation
reportlab==4.0.9

# ZIP file handling
zipfile36==0.1.3 